Move String code from main.zig to string.zig
This commit moves the source code of String to its
own separate file. This is nice because I now use
the entire file as the struct. Also, this would make
more sense as I am planning to add a StringUnmanaged
struct which to be similar to string but takes an
allocator every argument.