remove `heap`
Remove heap from the readme as in zig, the String can be stored anywhere, not necessarily on the heap. It may be stored on the stack with a FixedBufferAllocator. So, it is wrong to use heap in the readme as this makes it seem that the string can only be stored on the heap.